Finished Enemy Pathfinding! Made enemy class!
first I have added a js file that is an array of coordinates for the enemy path in stage 1, the enemies follow those points in the array. Made a class for the small enemy 64x64 pixels, it comes with x,y, and center positions and is constructed with those x and y positions. draw function draws the enemy, move function moves them along the path. I have tried adding many comments for clarification