Projectile Class, Vision Sphere, Firing!
Much has changed, firstly we have been having trouble with rotating our previous towers to point at the enemy they were firing at, so I have created a new tower sprite that does not rely on rotation to fire at the enemy. It will still have animation and I have created the sprite sheet for it already (it might need to be adjusted). The tower will fire when it has reached the top of it's animation (which right now doesn't exist) but for now I have it being shot at that exact frame of animation firing. The tower fire projectiles at enemies when they enter their range which is indicated as a circle that surrounds the tower, when the enemy enters it, the towers begin firing at the first one until it leaves it's range. Fire rate and speed of projectiles can be adjusted. Projectiles are removed when they collide with an enemy. Thats pretty much it. Made lots of comments