Game Terminate Condition, UI, Resources, More Sprites
So many changes, the player can now win or lose at the game. The player has lives that is decreased each time an enemy reaches the end. A wave system has been implemented, the player only wins if they survive the final wave. The enemies have health which is decreased when a tower projectile hits it. If the enemy health reaches zero or less it is removed from the game. Also, there is a coins resource that is dropped from a dying enemy, coins are added to the player running coins total and can only place a tower if they have sufficient coins to afford said tower. The UI has included all of these changes so that the player can keep track of coins, lives, enemy health etc. I have also added more tower sprites to the working directory but aren't seen in-game yet. Nothing is balanced yet, but starting to feel like a real game now :)