Multiple Towers, Sidebar, Sprite Animations
We have a working way to select different towers. the player will cycle through them with the 'a' and 'd' keys. They are able to know what tower they are cycling through due to the sidebar graphic which displays the tower name, damage, range, and cost. With that is the implementation of the 3 more towers which for right now only have different damage, fire rate and range. (they also look different) the enemy and towers all have their own sprite animations now. The game really looks much more alive because of it! Game mechanics still need balancing, but all in due time.